chore: allow right click
